requirements.txt
fastapi
uvicorn
httpx
google-adk
pydantic
python-dotenv